requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis==5.2.0
celery[redis]==5.4.0
openai==1.54.0
twilio==9.3.0
deepgram-sdk==3.7.0
elevenlabs==1.9.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
httpx==0.27.2
websockets==13.1
python-dotenv==1.0.1
cryptography==43.0.3
email-validator==2.2.0
stripe==11.1.0
